Add chrome_extension/bg.html and chrome_extension/js/notify.js
bg.html will run in the background - it will be the place to put any
scripts that need to run continuously. I added jquery.min.js, though I
don't know that it will be necessary or prudent. bg.html also includes
js/notify.js, which was created here. notify.js includes a global object
`Ins` (that should maybe be moved to it's own file?) and a function to
show notifications (as necessary) once per minute.